// source/frontend/src/models/api/generated/ApiGen_Concepts_LeaseTenant.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Base_BaseAudit } from './ApiGen_Base_BaseAudit';
import { ApiGen_Base_CodeType } from './ApiGen_Base_CodeType';
import { ApiGen_Concepts_Organization } from './ApiGen_Concepts_Organization';
import { ApiGen_Concepts_Person } from './ApiGen_Concepts_Person';

// LINK: @backend/apimodels/Models/Concepts/Lease/LeaseTenantModel.cs
export interface ApiGen_Concepts_LeaseTenant extends ApiGen_Base_BaseAudit {
  leaseTenantId: number | null;
  leaseId: number;
  personId: number | null;
  person: ApiGen_Concepts_Person | null;
  organizationId: number | null;
  organization: ApiGen_Concepts_Organization | null;
  note: string | null;
  primaryContact: ApiGen_Concepts_Person | null;
  primaryContactId: number | null;
  lessorType: ApiGen_Base_CodeType<string> | null;
  tenantTypeCode: ApiGen_Base_CodeType<string> | null;
}