// source/frontend/src/models/api/generated/ApiGen_Concepts_OrganizationPerson.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Base_BaseConcurrent } from './ApiGen_Base_BaseConcurrent';
import { ApiGen_Concepts_Person } from './ApiGen_Concepts_Person';

// LINK: @backend/apimodels/Models/Concepts/Organization/OrganizationPersonModel.cs
export interface ApiGen_Concepts_OrganizationPerson extends ApiGen_Base_BaseConcurrent {
  person: ApiGen_Concepts_Person | null;
  personId: number;
  organizationId: number;
}